+++ flutter-app/vault/lib/pages/SignUpForm.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:vault/ChangingText.dart';
-import 'package:vault/MainMenu.dart';
-import 'package:vault/ServerAPI.dart';
-import 'package:vault/SignInForm.dart';
+import 'package:vault/widgets/ChangingText.dart';
+import 'package:vault/pages/MainMenu.dart';
+import 'package:vault/utils/ServerAPI.dart';
+import 'package:vault/pages/SignInForm.dart';
 
 class SignUpForm extends StatefulWidget {
   const SignUpForm({Key key, this.serverAPI}) : super(key: key);
