+++ flutter-app/vault/lib/utils/CryptoTools.dart
